+++ src/words/ff_words_comp_p.h
@@ -9,7 +9,7 @@
 case FF_OP_DOES_RUNTIME:
     {
         ff_word_t *nw = (ff_word_t *)(intptr_t)*ip++;
-        _FF_RSO(1);
+        _FF_RSO_T(1);
         _FF_SO(1);
         if (ff->state & FF_STATE_BACKTRACE)
             ff_bt_stack_push(BT, ff->cur_word);
@@ -164,7 +164,7 @@ case FF_OP_EXECUTE:
  * then bail out of the definition like an EXIT.
  */
 case FF_OP_DOES:
-    _FF_RSL(1);
+    _FF_RSL_T(1);
     ff_dict_top(&ff->dict)->does = ip;
     ff_dict_top(&ff->dict)->opcode = FF_OP_DOES_RUNTIME;
     /* Simulate EXIT to bail out of the definition. */